Choosing any rectangular Korean data file using interactive GUI dialog box.

The basic functionality is exactly the same as pick, but optimized for Korean R users.

picko(file = NULL, ...)

Arguments

file

Either a path to a file, a connection, or literal data (either a single string or a raw vector). The default is NULL, which pops up an interactive GUI file choose dialogue box for users unless an explicit path/to/filename is given. Each corresponding function depending upon a file extension will be automatically matched and applied once you pick up your file using either the GUI-file-chooser dialog box or explicit path/to/filename.

...

Any additional arguments available for each file type and extension: vroom for 'CSV' (Comma-Separated Values); 'CSV2' (Semicolon-Separated Values); 'TSV' (Tab-Separated Values); 'txt' (plain text) files; read_excel for 'Excel' files; read_spss for 'SPSS' files; read_stata for 'Stata' files; read_sas for 'SAS' files; read_document for 'Microsoft Word', 'PDF', 'RTF', 'HTML', 'HTM', and 'PHP' files; fromJSON for 'JSON' files; read_mbox for 'mbox' files; render for 'Rmd' files; source for 'R' files; readRDS for 'RDS' files; load for 'RDA' and 'RDATA' files.

Value

tibble (table data.frame) object of the chosen rectangular data file will be returned.

Details

picko

See example below.

See also

Author

JooYoung Seo, jooyoung@psu.edu

Soyoung Choi, sxc940@psu.edu

Examples

# Choosing file and saving it into a variable
## Scenario 1: Picking up a file using interactive GUI dialog box:
if (interactive()) {
  library(ezpickr)
  data <- picko()
}

## Scenario 2: Picking up a file using an explicit file name ("test.sav" in the example below;
## however, you can feed other files through this function
## such as *.SAS, *.DTA, *.csv, *.csv2, *.tsv, *.xlsx, *.txt,
## *.html, webpage URL, *doc, *.docx, *.pdf, *.rtf, *.json, *.Rda, *.Rdata, and more):
if (interactive()) {
  library(ezpickr)
  test <- system.file("extdata", "airquality.sav", package = "ezpickr")
  data <- picko(test)

  # Now you can use the imported file as a tibble.
  str(data)
}