Seamlessly manipulate any rectangular data file between an Excel window and R session.

You can use this function for loading and manipulating any data.frame, data_frame, tbl_df, matrix, table, vector, or DocumentTermMatrix objects into your system-default spreadsheet software (e.g., Excel) in a real time. This function has employed write_xlsx under the hood.

viewxl(x, ...)

Arguments

x

An object of class data.frame, matrix, table, vector, or DocumentTermMatrix.

...

Any additional arguments available for write_xlsx.

Value

Data object opened in a preferable spreadsheet application window which will in turn be called on your R session again.

Details

viewxl

See example below.

Author

JooYoung Seo, jooyoung@psu.edu

Soyoung Choi, sxc940@psu.edu

Examples

if (interactive()) {
  library(ezpickr)
  data(airquality)
  str(airquality)

  ## View your data object in your spreadsheet software:
  viewxl(airquality)
  # Then, when necessary, you can modify the opened data
  # in the spreadsheet and save it as a new data.

  # You can pass a list object to the `view()` function like below:
  l <- list(iris = iris, mtcars = mtcars, chickwts = chickwts, quakes = quakes)
  viewxl(l)
  # Then, each list item will appear in your Excel window sheet by sheet.
}