Choosing any rectangular data file using interactive GUI dialog box.

You can alternatively use this function for choosing *.csv, *.csv2, *.tsv, *.txt, *.xls, *.xlsx, *.json, *.html, *.htm, *.php, *.pdf, *.doc, *.docx, *.rtf, *.RData, *.Rda, *.RDS, *.sav (SPSS), *.por, *.sas7bdat, *.sas7bcat, and *.dta files in an interactive GUI mode A file choose dialog box will be prompted.

pick(file = NULL, ...)

Arguments

file
Either a path to a file, a connection, or literal data (either a single string or a raw vector). The default is NULL, which pops up an interactive GUI file choose dialogue box for users unless an explicit path/to/filename is given. Each corresponding function depending upon a file extension will be automatically matched and applied once you pick up your file using either the GUI-file-chooser dialog box or explicit path/to/filename.

...
Any additional arguments available for each file type and extension: read_csv for CSV (Comma-Separated Values) files; read_csv2 for CSV2 (Semicolon-Separated Values) files; read_tsv for 'TSV' (Tab-Separated Values) files; read_file for 'txt' (plain text) files; read_excel for 'Excel' files; read_spss for 'SPSS' files; read_stata for 'Stata' files; read_sas for 'SAS' files; read_document for 'Microsoft Word', 'PDF', 'RTF', 'HTML', 'HTM', and 'PHP' files; fromJSON for 'JSON' files; readRDS for 'RDS' files; load for 'RDA' and 'RDATA' files.

Value

tibble (table data.frame) object of the chosen rectangular data file will be returned.

Details

pick

See example below.

Examples

# NOT RUN {
library(ezpickr)
# Choosing file and saving it into a variable
## Scenario 1: Picking up a file using interactive GUI dialog box:
data <- pick()

## Scenario 2: Picking up a file using an explicit file name ("test.sav" in the example below;
## however, you can feed other files through this function
## such as *.SAS, *.DTA, *.csv, *.csv2, *.tsv, *.xlsx, *.txt,
## *.html, webpage URL, *doc, *.docx, *.pdf, *.rtf, *.json, *.Rda, *.Rdata, and more):
data <- pick("test.sav")

# Now you can use the imported file as a data frame.
str(data)
# }